+++ docs/代办列表.md
@@ -13,9 +13,10 @@
 | 参数配置：用于系统内的一些全局参数设置 （参考：https://saber.bladex.vip/#/system/param） | - | 
 | 基于登录日志，做一些用户分析（用户地图等） | - | 
 | 代码生成器 初始化模块 & 前端页面生成 | - | 
-| 远程查询自动注入（字典&人员信息等） | - | 
+| 远程查询自动注入（字典&人员信息等） | ok | 
 | 首页右上角显示消息列表 | - | 
 | 规划如何将 zuihou-commons 模块独立 | - | 
+| 第三方登录 | - | 
 
 
 ## 2019年09月06日23:57:21
