+++ docs/升级日志.md
@@ -1,10 +1,20 @@
-# 升级日志
+#升级日志 2020年01月28日17:42:42
+1. 修复 bug    <a href="https://github.com/zuihou/zuihou-ui/issues/3"># fix 3</a> 
+2. 升级SpringBoot 版本至 2.2.4
+3. 升级MybatisPlus 版本至 3.3.1
+4. 升级spring  版本至 5.2.3
+5. 代码生成器支持 生成父类为TreeEntity类的结构
+6. 代码生成器支持 解析数据库注释中的@InjectionField类型
+7. 数据注入工具类重命名为： zuihou-injection-starter
+
+
+# 升级日志 2020年01月21日
 0，定时任务 增加重置演示环境数据功能
 1，完善地区相关接口，并优化地区表结构
 2，对接 地区 相关前端页面
 3，新增 数据字典相关方法
 4，新增 爬虫，用于从国家统计局爬取全国的省市区数据
-5，新增 数据聚合工具（zuihou-remote-data-starter), 用于解决跨库、跨表、跨服务的 关联数据 回显之痛。支持对静态数据属性(数据字典)、动态主键数据等进行自动注入和转化。
+5，新增 数据聚合工具（zuihou-injection-starter), 用于解决跨库、跨表、跨服务的 关联数据 回显之痛。支持对静态数据属性(数据字典)、动态主键数据等进行自动注入和转化。
 6，新增 树形基础实体TreeEntity， 使得实体支持树状结构
 7，新增 Mybatis自定义类型处理器，使得Mybatis 能够将数据库中的字段解析成Java中的自定义类型
