+++ zuihou-backend/zuihou-authority/zuihou-authority-biz/src/main/java/com/github/zuihou/authority/service/core/OrgService.java
@@ -1,9 +1,12 @@
 package com.github.zuihou.authority.service.core;
 
-import java.util.List;
-
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.github.zuihou.authority.entity.core.Org;
+
+import java.io.Serializable;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * <p>
@@ -30,4 +33,12 @@ public interface OrgService extends IService<Org> {
      * @return
      */
     boolean remove(List<Long> ids);
+
+    /**
+     * 根据 id 查询组织，并转换成Map结构
+     *
+     * @param ids
+     * @return
+     */
+    Map<Serializable, Object> findOrgByIds(Set<Serializable> ids);
 }
