+++ zuihou-backend/zuihou-authority/zuihou-authority-biz/src/main/java/com/github/zuihou/authority/service/core/impl/OrgServiceImpl.java
@@ -7,13 +7,19 @@ import com.github.zuihou.authority.entity.core.Org;
 import com.github.zuihou.authority.service.auth.RoleOrgService;
 import com.github.zuihou.authority.service.core.OrgService;
 import com.github.zuihou.database.mybatis.conditions.Wraps;
+import com.github.zuihou.database.mybatis.conditions.query.LbqWrapper;
+import com.github.zuihou.utils.MapHelper;
+import com.google.common.collect.ImmutableMap;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.Serializable;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -49,4 +55,23 @@ public class OrgServiceImpl extends ServiceImpl<OrgMapper, Org> implements OrgSe
         List<Long> idList = list.stream().mapToLong(Org::getId).boxed().collect(Collectors.toList());
         return !idList.isEmpty() ? super.removeByIds(idList) : true;
     }
+
+    /**
+     * TODO 这里应该做缓存
+     *
+     * @param ids
+     * @return
+     */
+    @Override
+    public Map<Serializable, Object> findOrgByIds(Set<Serializable> ids) {
+        LbqWrapper<Org> query = Wraps.<Org>lbQ()
+                .in(Org::getId, ids)
+                .eq(Org::getStatus, true);
+        List<Org> list = super.list(query);
+
+        //key 是 组织id， value 是org 对象
+        ImmutableMap<Serializable, Object> typeMap = MapHelper.uniqueIndex(list, Org::getId, (org) -> org);
+        return typeMap;
+    }
+
 }
