+++ zuihou-backend/zuihou-authority/zuihou-authority-entity/pom.xml
@@ -28,7 +28,7 @@
         </dependency>
         <dependency>
             <groupId>com.github.zuihou</groupId>
-            <artifactId>zuihou-remote-data-starter</artifactId>
+            <artifactId>zuihou-injection-starter</artifactId>
             <version>${commons.version}</version>
         </dependency>
         <!--        <dependency>-->
