+++ zuihou-backend/zuihou-authority/zuihou-authority-entity/src/main/java/com/github/zuihou/authority/dto/core/StationSaveDTO.java
@@ -1,17 +1,13 @@
 package com.github.zuihou.authority.dto.core;
 
-import java.io.Serializable;
-
+import com.github.zuihou.model.RemoteData;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
+import lombok.*;
 import lombok.experimental.Accessors;
 import org.hibernate.validator.constraints.Length;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -45,7 +41,7 @@ public class StationSaveDTO implements Serializable {
      * #c_core_org
      */
     @ApiModelProperty(value = "组织ID")
-    private Long orgId;
+    private RemoteData orgId;
     /**
      * 状态
      */
