+++ zuihou-backend/zuihou-authority/zuihou-authority-server/src/main/resources/application-dev.yml
@@ -1,5 +1,5 @@
 zuihou:
-  remote:
+  injection:
     # 是否启用 远程数据 手动注入
     enabled: true
     # 是否启用 远程数据 注解注入
