--- zuihou-commons/zuihou-common/src/main/java/com/github/zuihou/common/constant/RemoteDataConstants.java
@@ -1,16 +1,0 @@
-package com.github.zuihou.common.constant;
-
-/**
- * 仅仅用于记录 RemoteField 注解相关的 接口和方法名称
- * <p>
- * 切记，该类下的接口和方法，一定要自己手动创建，否则会注入失败
- *
- * @author zuihou
- * @date 2020年01月20日11:16:37
- */
-public abstract class RemoteDataConstants {
-
-    public static final String DICTIONARY_ITEM_CLASS = "dictionaryItemServiceImpl";
-
-    public static final String DICTIONARY_ITEM_METHOD = "findDictionaryItem";
-}
