+++ zuihou-commons/zuihou-config/src/main/resources/common.yml
@@ -55,8 +55,8 @@ spring:
 
 dozer:
   mappingFiles:
-    - classpath:dozer/global.dozer.xml
-    - classpath:dozer/biz.dozer.xml
+    - classpath*:dozer/*.dozer.xml
+
 management:
   endpoints:
     web:
