--- zuihou-commons/zuihou-dozer-starter/src/main/resources/dozer/biz.dozer.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mappings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xmlns="http://dozermapper.github.io/schema/bean-mapping"
-          xsi:schemaLocation="http://dozermapper.github.io/schema/bean-mapping http://dozermapper.github.io/schema/bean-mapping.xsd">
-
-    <mapping date-format="yyyy-MM-dd HH:mm:ss">
-        <class-a>com.github.zuihou.authority.entity.auth.Menu</class-a>
-        <class-b>com.github.zuihou.authority.dto.auth.VueRouter</class-b>
-        <field>
-            <a>name</a>
-            <b>meta.title</b>
-        </field>
-        <field>
-            <a>name</a>
-            <b>name</b>
-        </field>
-        <field>
-            <a>icon</a>
-            <b>meta.icon</b>
-        </field>
-    </mapping>
-
-
-</mappings>
