+++ zuihou-commons/zuihou-injection-starter/pom.xml
@@ -6,40 +6,31 @@
         <artifactId>zuihou-commons</artifactId>
         <groupId>com.github.zuihou</groupId>
         <version>1.0-SNAPSHOT</version>
-        <relativePath>../</relativePath>
     </parent>
-
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>zuihou-dozer-starter</artifactId>
-    <name>${project.artifactId}</name>
-    <description>优雅bean转换模块</description>
+
+    <artifactId>zuihou-injection-starter</artifactId>
 
     <dependencies>
-        <!--        <dependency>-->
-        <!--            <groupId>com.github.dozermapper</groupId>-->
-        <!--            <artifactId>dozer-spring-boot-starter</artifactId>-->
-        <!--        </dependency>-->
-
+        <dependency>
+            <artifactId>zuihou-core</artifactId>
+            <groupId>com.github.zuihou</groupId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-aop</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-autoconfigure</artifactId>
+            <scope>compile</scope>
         </dependency>
-
-        <dependency>
-            <groupId>com.github.dozermapper</groupId>
-            <artifactId>dozer-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.github.dozermapper</groupId>
-            <artifactId>dozer-spring4</artifactId>
-        </dependency>
-
-        <!-- 必须有这个jar 才能提示 -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-configuration-processor</artifactId>
             <optional>true</optional>
             <scope>compile</scope>
         </dependency>
+
     </dependencies>
 </project>
