+++ zuihou-commons/zuihou-injection-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    com.github.zuihou.injection.InjectionDataAutoConfiguration
