--- zuihou-commons/zuihou-remote-data-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    com.github.zuihou.remotedata.RemoteDataAutoConfiguration
